Optimize shrink, rewrite, and polish all code.  V2
* Added debug option to build options.
should be made as release x64 anyway.

* Added auto zip ready dll to Build folder

* shrinked code to use simple groups lists

* included also grids connected by landing gear to main grid. including all checks on them.
Owner is checked on all grids exept subgrids.
So connected by landing gear are checked for owner.

If there is pilot in any grid, and he is online, 
code will check if player is faction member, or in peace with your faction. if not of dont have faction, then we get alert to remove players from blocks.

if freindly! then he will be removed from block, and grid will be fixed.
All offline players will remain in blocks.
 
* Used parallel foreach